feat(wintertc): migrate queueMicrotask from boa_runtime
Move the queueMicrotask implementation from boa_runtime into boa_wintertc,
following the move-and-re-export pattern used for atob/btoa (#5418).

- core/runtime/src/microtask is deleted; the code now lives only in
  boa_wintertc.
- boa_runtime re-exports it (pub use boa_wintertc::microtask) so the public
  path boa_runtime::microtask::register is unchanged.
- MicrotaskExtension delegates to boa_wintertc::microtask::register.

The unit test is rewritten to record execution order in a global array
rather than through console, since console is not part of boa_wintertc's
test surface (it is still a stub pending its own migration).
